changed migrations

+++ db/migrations/20260524104410_add_visits_table.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 CREATE TABLE IF NOT EXISTS link_visits (
 	id BIGSERIAL PRIMARY KEY,
-	link_id INT,
+	link_id BIGINT,
 	ip VARCHAR(45),
 	user_agent VARCHAR(255),
 	referer TEXT,
